add todo crud

+++ src/main/java/com/example/todolist/repository/ToDoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
+public interface ToDoRepository extends JpaRepository<ToDo, Long> {
     ToDo findById(int id);
     ToDo findByStatus(ToDoStatus status);
 }
